changed to components - initial commit
