+++ src/Header/Header.tsx
@@ -1,0 +1,9 @@
+const Header = () => {
+  return (
+    <header className='cv__header'>
+      <h1>CV</h1>
+    </header>
+  );
+};
+
+export { Header };
